+++ src/main/java/ces/springmvc/bookmanagementspringmvc/dao/Book_BookTypeDAO.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import ces.springmvc.bookmanagementspringmvc.entity.Book_BookTypeEntity;
 
 @Repository
-public interface Book_BookTypeDAO extends JpaRepository<Book_BookTypeEntity, Integer>{
+public interface Book_BookTypeDAO extends JpaRepository<Book_BookTypeEntity, Long>{
 
 }
